factor out to vote_graph implementation and generic vote type
